+++ src/main/java/com/youthcase/orderflow/stk/service/STKServiceImpl.java
@@ -126,7 +126,7 @@ public class STKServiceImpl implements STKService {
     @Override
     public List<STK> getStockByProductGtin(String gtin) {
         // 재고가 0보다 큰 활성 재고 랏 목록을 유통기한 순으로 조회
-        return stkRepository.findByProduct_GtinAndQuantityGreaterThanOrderByLot_ExpDateAsc(gtin, 0);
+        return stkRepository.findByProduct_GtinAndQuantityGreaterThanOrderByLot_ExpDateAsc(gtin, 0L);
     }
 
     @Override
@@ -172,7 +172,7 @@ public class STKServiceImpl implements STKService {
     public List<STK> disposeExpiredStock(LocalDate targetDate) {
         List<STK> expiredStocks = stkRepository.findExpiredActiveStockBefore(targetDate);
         for (STK stock : expiredStocks) {
-            stock.setQuantity(0);
+            stock.setQuantity(0L);
             stock.updateStatus("DISPOSED");
             stkRepository.save(stock);
         }
@@ -196,23 +196,23 @@ public class STKServiceImpl implements STKService {
         // ... (출고 차감 로직 생략 없이 유지)
         for (StockDeductionRequestDTO.DeductionItem item : requestDTO.getItems()) {
             String gtin = item.getGtin();
-            Integer requiredQuantity = item.getQuantity();
+            Long requiredQuantity = item.getQuantity();
 
-            List<STK> fifoStocks = stkRepository.findByProduct_GtinAndQuantityGreaterThanOrderByLot_ExpDateAsc(gtin, 0);
+            List<STK> fifoStocks = stkRepository.findByProduct_GtinAndQuantityGreaterThanOrderByLot_ExpDateAsc(gtin, 0L);
 
-            int remainingToDeduct = requiredQuantity;
+            Long remainingToDeduct = requiredQuantity;
 
             for (STK stock : fifoStocks) {
                 if (remainingToDeduct <= 0) break;
 
-                int stockQuantity = stock.getQuantity();
+                Long stockQuantity = stock.getQuantity();
 
                 if (stockQuantity >= remainingToDeduct) {
                     stock.setQuantity(stockQuantity - remainingToDeduct);
-                    remainingToDeduct = 0;
+                    remainingToDeduct = 0L;
                 } else {
                     remainingToDeduct -= stockQuantity;
-                    stock.setQuantity(0);
+                    stock.setQuantity(0L);
                     stock.updateStatus("INACTIVE");
                 }
 
@@ -232,7 +232,7 @@ public class STKServiceImpl implements STKService {
     @Override
     public STK findFirstAvailableByGtin(String gtin) {
         return stkRepository
-                .findByProduct_GtinAndQuantityGreaterThanOrderByLot_ExpDateAsc(gtin, 0)
+                .findByProduct_GtinAndQuantityGreaterThanOrderByLot_ExpDateAsc(gtin, 0L)
                 .stream()
                 .findFirst()
                 .orElseThrow(() -> new NoSuchElementException("해당 상품의 재고가 없습니다."));
@@ -248,21 +248,21 @@ public class STKServiceImpl implements STKService {
             int requestedQuantity = item.getQuantity();
 
             // Lot ID로 활성 재고를 찾습니다.
-            Optional<STK> stkOptional = stkRepository.findByLot_LotIdAndQuantityGreaterThan(lotId, 0);
+            Optional<STK> stkOptional = stkRepository.findByLot_LotIdAndQuantityGreaterThan(lotId, 0L);
 
             if (stkOptional.isEmpty()) {
                 throw new NoSuchElementException("Lot ID " + lotId + "에 해당하는 활성 재고를 찾을 수 없습니다.");
             }
 
             STK stock = stkOptional.get();
-            int currentQuantity = stock.getQuantity();
+            Long currentQuantity = stock.getQuantity();
 
             if (requestedQuantity <= 0 || requestedQuantity > currentQuantity) {
                 throw new IllegalArgumentException("Lot ID " + lotId + "에 대한 폐기 요청 수량(" + requestedQuantity + ")이 유효하지 않습니다.");
             }
 
             // 재고 수량 감소 및 상태 변경
-            int newQuantity = currentQuantity - requestedQuantity;
+            Long newQuantity = currentQuantity - requestedQuantity;
             stock.setQuantity(newQuantity);
 
             if (newQuantity == 0) {
@@ -291,7 +291,7 @@ public class STKServiceImpl implements STKService {
         for (AdjustmentRequest.AdjustmentItem item : request.getItems()) {
 
             Long lotId = item.getLotId();
-            int targetQuantity = item.getTargetQuantity();
+            Long targetQuantity = item.getTargetQuantity();
 
             // Lot ID로 조정할 STK 재고 항목을 조회합니다. (재고가 0 이하라도 조회되어야 하므로 findByLot_LotId 사용)
             Optional<STK> stkOptional = stkRepository.findByLot_LotId(lotId);
@@ -326,7 +326,7 @@ public class STKServiceImpl implements STKService {
     @Override
     public List<STK> findStocksRequiringAdjustment() {
         // 0 이하의 수량을 가진 재고를 조회합니다.
-        return stkRepository.findByQuantityLessThanEqual(0);
+        return stkRepository.findByQuantityLessThanEqual(0L);
     }
 
     //GR
@@ -348,7 +348,7 @@ public class STKServiceImpl implements STKService {
         if (existingOpt.isPresent()) {
             // ✅ 기존 재고가 있으면 수량만 증가
             stk = existingOpt.get();
-            int newQty = stk.getQuantity() + qty.intValue();
+            Long newQty = stk.getQuantity() + qty;
             stk.setQuantity(newQty);
             stk.setLastUpdatedAt(LocalDateTime.now());
         } else {
@@ -358,7 +358,7 @@ public class STKServiceImpl implements STKService {
                     .product(product)
                     .lot(lot)
                     .goodsReceipt(null)
-                    .quantity(qty.intValue())
+                    .quantity(qty)
                     .hasExpirationDate(expDate != null)
                     .status("ACTIVE")
                     .lastUpdatedAt(LocalDateTime.now())
@@ -377,7 +377,7 @@ public class STKServiceImpl implements STKService {
         STK stk = stkRepository.findByWarehouseAndProductAndLot(warehouseId, gtin, lotNo)
                 .orElseThrow(() -> new IllegalArgumentException("해당 재고 없음"));
 
-        int remain = stk.getQuantity() - qty.intValue();
+        Long remain = stk.getQuantity() - qty;
         if (remain < 0) {
             throw new IllegalStateException("재고 수량 부족: " + stk.getProductName());
         }
